64-bit addresses and sizes (draft)

+++ board/rcm/common/flw_dev_list.h
@@ -10,9 +10,9 @@ enum FLW_DEV_TYPE {FLWDT_UNKNOWN=0, FLWDT_SF, FLWDT_MMC, FLWDT_NAND, FLWDT_NOR};
 
 struct flw_dev_t;
 
-typedef int (*flw_erase)(struct flw_dev_t* fd, unsigned long addr, unsigned long size);
-typedef int (*flw_write)(struct flw_dev_t* fd, unsigned long addr, unsigned long size, const char* data);
-typedef int (*flw_read)(struct flw_dev_t* fd, unsigned long addr, unsigned long size, char* data);
+typedef int (*flw_erase)(struct flw_dev_t* fd, unsigned long long addr, unsigned long long size);
+typedef int (*flw_write)(struct flw_dev_t* fd, unsigned long long addr, unsigned long long size, const char* data);
+typedef int (*flw_read)(struct flw_dev_t* fd, unsigned long long addr, unsigned long long size, char* data);
 
 struct flw_dev_info_t
 {
